requirements/base.txt
Django==4.2.13
djangorestframework==3.15.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
drf-spectacular==0.27.2
anthropic==0.28.0
psycopg2-binary==2.9.9
django-environ==0.11.2
requests==2.32.3
Pillow==10.3.0
